tale-generator/pyproject.toml
[project]
name = "tale-generator"
version = "0.1.0"
description = "API service for generating children's bedtime stories"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "openai>=1.54.0",
    "supabase>=2.24.0",
    "openai-cost-calculator>=1.1.1",
    "requests>=2.32.5",
    "elevenlabs>=2.24.0",
    "click>=8.1.0",
    "tabulate>=0.9.0",
    "redis>=5.0.0",
    "pydantic-settings>=2.12.0",
    "langgraph>=1.0.4",
    "langchain-core>=1.1.1",
    "langchain-openai>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pyyaml>=6.0.3",
]

[project.scripts]
start = "main:run"
